Change Mysql to H2 and column for function

+++ src/main/java/es/upm/fi/dia/oeg/rdb/DatasetImport.java
@@ -1,17 +1,17 @@
 package es.upm.fi.dia.oeg.rdb;
 
-import es.upm.fi.dia.oeg.Utils;
+
 import org.apache.commons.io.FileUtils;
-import org.apache.commons.io.IOUtils;
+import org.apache.jena.atlas.json.JsonObject;
 import org.json.JSONArray;
 import org.json.JSONObject;
+import org.json.JSONString;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
-import java.io.FileReader;
 import java.net.URL;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
@@ -28,20 +28,37 @@ public class DatasetImport {
                 String relativePath= "datasets/"+dataset.getString("databaseName");
                 File f = new File(relativePath);
                 f.mkdir();
-                if (dataset.getString("url").matches("http.*")) {
-                    f = new File(relativePath+"/dataset."+dataset.getString("compression"));
-                    FileUtils.copyURLToFile(new URL(dataset.getString("url")), f);
-                }
-                else {
-                    FileUtils.copyDirectory(new File(dataset.getString("url")), f);
-                }
+                Object source = dataset.get("source");
+                if(source instanceof String) {
+                    String source_string = (String) source;
+                    if (source_string.matches("http.*")) {
+                        f = new File(relativePath + "/dataset." + dataset.getString("compression"));
+                        FileUtils.copyURLToFile(new URL(source_string), f);
+                    }
+                    else {
+                        FileUtils.copyDirectory(new File(source_string), f);
+                    }
 
-                if(dataset.getString("compression").equals("zip")){
-                    unzip(relativePath,f.getName());
-                    f.delete();
+                    if(dataset.getString("compression").equals("zip")){
+                        unzip(relativePath,f.getName());
+                        f.delete();
+                    }
+                    else if(!dataset.getString("compression").isEmpty()){
+                        _log.error("The compression of "+dataset.getString("databaseName")+" CSV set should be ZIP");
+                    }
                 }
-                else if(!dataset.getString("compression").isEmpty()){
-                    _log.error("The compression of "+dataset.getString("databaseName")+" CSV set should be ZIP");
+                else if (source instanceof JSONArray){
+                    JSONArray source_array = (JSONArray) source;
+                    for (Object s : source_array){
+                        String path = (String) s;
+                        String file_name = path.split("\\.csv")[0].split("/")[path.split("\\.csv")[0].split("/").length-1];
+                        if(path.matches("http.*")){
+                            FileUtils.copyURLToFile(new URL(path),new File(relativePath+"/"+file_name+".csv"));
+                        }
+                        else{
+                            FileUtils.copyFile(new File(path),new File(relativePath+"/"+file_name+".csv"));
+                        }
+                    }
                 }
             }
 
@@ -54,11 +71,13 @@ public class DatasetImport {
     private static void unzip (String path, String name){
         try {
             byte[] buffer = new byte[1024];
+            File f = new File(path+"/data");
+            f.mkdir();
             ZipInputStream zis = new ZipInputStream(new FileInputStream(path+"/"+name));
             ZipEntry zipEntry = zis.getNextEntry();
             while (zipEntry != null) {
                 String fileName = zipEntry.getName();
-                File newFile = new File(path+"/"+fileName);
+                File newFile = new File(path+"/data/"+fileName.substring(0,fileName.lastIndexOf("."))+".csv");
                 FileOutputStream fos = new FileOutputStream(newFile);
                 int len;
                 while ((len = zis.read(buffer)) > 0) {
